+++ src/helper/index.js
@@ -1,12 +1,16 @@
 const { getActiveNetwork , getTransactionVersion } = require('./getNetwork')
 const { generateUnsignedTransaction } = require('./utils/generateUnsignedTransaction')
 const { generatePostConditions } = require('./utils/generatePostConditions')
-const { gerenateFunctionArgs } = require('./utils/gerenateFunctionArgs')
+const { generateFunctionArgs } = require('./utils/generateFunctionArgs');
+const { generatePayload } = require('./utils/generatePaylod');
+const { generateStacksTransactionObject } = require('./utils/generateStacksTransactionObject');
 
 module.exports = {
     getActiveNetwork,
     getTransactionVersion,
     generateUnsignedTransaction,
     generatePostConditions,
-    gerenateFunctionArgs
+    generateFunctionArgs,
+    generatePayload,
+    generateStacksTransactionObject
 }